Move files to traits
