+++ java/toroni/src/test/java/toroni/traits/posix/PthreadRobustMutexTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package toroni.traits;
+package toroni.traits.posix;
 
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
